requirements.txt
appdirs==1.4.4
astroid==2.11.2
attrs==21.4.0
backports.entry-points-selectable==1.1.1
black==22.3.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
colorzero==2.0
coverage==6.3.2
dill==0.3.4
distlib==0.3.4
filelock==3.6.0
gpiozero==1.6.2
identify==2.4.12
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy==0.942
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
pigpio==1.78
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.18.1
py==1.11.0
pylint==2.13.5
pyparsing==3.0.8
pyserial==3.5
PySimpleGUI==4.59.0
pytest==7.1.1
PyYAML==6.0
qrcode==7.3.1
regex==2022.3.15
requests==2.27.1
requests-aws4auth==1.1.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
transitions==0.8.11
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
wrapt==1.14.0
RPi.GPIO==0.7.1
